Add TeleOp Stubs and Control of Wheels
The controls of the wheels aren't fully debugged. They work mostly in
the four cardinal directions and sometimes diagonally, but it is not
doing all angles smoothly.

Added the tele op stubs for each button and split them up into different
mechanisms on the robot. Included comments and some TODOs of ideas we
need to do.