Test Motors with New Names
Create a wormDriveMotor and an intakeMotor.  Made actual spring position
values to test out the wormDriveMotor.  Add new motors to the
MotorEnumeration.  Comment a lot of arm code, we need to figure out what
to keep.  Add spring sensor that hasn't physically been added to the
robot yet.
Write code for tele-Op button pressing.  This is for testing out each
spring position.