Started to Implement a Top Down Library Function
Organized the init method in a logical order.
Created a servoArray to express all servos.
Combined the runRobotToPositionFB/SidewaysRight/SidewaysLeft to use a
switch for direction.
Did the same for moveRobotFB/SidewaysRight/SidewaysLeft.
Combined all of the set and get servos into two main methods using
switches depending on what servo is inputted.