Created Timer Array for Generic and DPAD Timer
Timer Array is used to implement more than one timer making it easier to
add timers.

DPAD Timer is used to time the long press of the dpad buttons in order
to make the robot move sideways left to score the beacons. However, the
robot does not run as planned and does random turns and jerky movements.
We hypothesized that the dpad aren't meant to be held because the dpad
reads more than one button after being held for a while (i.e. when the
up is held, sometimes the left or right will be read).

Solutions:
Have a separate button that just moves sideways left.
Press the button twice in order to move forward and a third time to
stop.